pyproject.toml
[project]
name = "t1-construcao"
version = "0.1.0"
description = ""
authors = [
    {name = "Pedro"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "msgspec (>=0.19.0,<0.20.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "aerich (>=0.9.1,<0.10.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "t1_construcao", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pylint = "^3.3.7"
pyright = "^1.1.403"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

[tool.pyright]
include = ["src", "tests"]
exclude = [
    "**/__pycache__",
    "**/node_modules",
    "**/.venv",
    "**/venv",
    "**/build",
    "**/dist",
    "**/.git"
]
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedVariable = true
reportDuplicateImport = true
reportIncompatibleVariableOverride = "none"
pythonVersion = "3.12"
pythonPlatform = "Linux"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "src.t1_construcao.infrastructure._tortoise_config.TORTOISE_ORM"
location = "./src/t1_construcao/infrastructure/migrations"
src_folder = "./."

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-class-docstring", 
    "missing-function-docstring",
    "too-few-public-methods",
    "unnecessary-ellipsis",
    "duplicate-code",
]

[tool.pylint.master]
ignore-paths = [
    "src/t1_construcao/infrastructure/migrations",
    ".*/__pycache__/.*",
    ".*/migrations/.*"
]

[tool.pylint.format]
max-line-length = 100

[tool.pylint.basic]
good-names = ["i", "j", "k", "ex", "Run", "_", "id", "db"]

[tool.pylint.design]
max-args = 10
max-locals = 20
max-returns = 6
max-branches = 15
